Break a few more circular dependencies
This is done by breaking the `Metrics` interface out of
the metrics index file. A similar approach was taken for the
ping payload.
