+++ glean/src/core/metrics/database.ts
@@ -8,12 +8,7 @@ import { Metric } from "./metric.js";
 import { createMetric, validateMetricInternalRepresentation } from "./utils.js";
 import { isObject, isUndefined, JSONObject, JSONValue } from "../utils.js";
 import { StorageBuilder } from "../../platform/index.js";
-
-export interface Metrics {
-  [aMetricType: string]: {
-    [aMetricIdentifier: string]: JSONValue
-  }
-}
+import { Metrics } from "./metrics_interface";
 
 /**
  * Verifies if a given value is a valid Metrics object.
